Adding Universal Windows Store projects/solutions, and ICoreWindow support.

+++ src/libGLESv2/renderer/d3d11/InputLayoutCache.cpp
@@ -185,8 +185,8 @@ GLenum InputLayoutCache::applyVertexBuffers(TranslatedAttribute attributes[gl::M
     }
 
     bool dirtyBuffers = false;
-    size_t minDiff = gl::MAX_VERTEX_ATTRIBS;
-    size_t maxDiff = 0;
+    UINT minDiff = gl::MAX_VERTEX_ATTRIBS;
+    UINT maxDiff = 0;
     for (unsigned int i = 0; i < gl::MAX_VERTEX_ATTRIBS; i++)
     {
         ID3D11Buffer *buffer = NULL;
